// cosmowcl/src/app/shared/models/api/cosmoecm/creaDocumentiRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * documento
 * API per la gestione dei documenti
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreaDocumentoRequest } from './creaDocumentoRequest.d';


/**
 * Model che contiene i campi relativi ai documenti da creare
 */
export interface CreaDocumentiRequest { 
    documenti?: Array<CreaDocumentoRequest>;
}