// cosmowcl/src/app/shared/models/api/cosmoecm/documento.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * firma.v1
 * API per la gestione della firma digitale
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TipoDocumento } from './tipoDocumento.d';
import { ApprovazioneDocumento } from './approvazioneDocumento.d';
import { SmistamentoDocumento } from './smistamentoDocumento.d';
import { StatoDocumento } from './statoDocumento.d';
import { ContenutoDocumento } from './contenutoDocumento.d';
import { SigilloDocumento } from './sigilloDocumento.d';


/**
 * Model che contiene i campi relativi al documento da creare
 */
export interface Documento { 
    idPratica: number;
    idDocumentoExt?: string;
    titolo?: string;
    tipo?: TipoDocumento;
    descrizione?: string;
    idDocumentoParentExt?: string;
    id?: number;
    autore?: string;
    stato?: StatoDocumento;
    contenuti?: Array<ContenutoDocumento>;
    ultimaModifica?: string;
    smistamento?: SmistamentoDocumento;
    allegati?: Array<Documento>;
    approvazioni?: Array<ApprovazioneDocumento>;
    sigillo?: Array<SigilloDocumento>;
}